+++ fxn/cli/__init__.py
@@ -9,7 +9,7 @@ from .auth import app as auth_app
 from .compile import compile_predictor
 from .misc import cli_options
 from .predictions import create_prediction
-from .predictors import retrieve_predictor
+from .predictors import archive_predictor, delete_predictor, retrieve_predictor
 from ..version import __version__
 
 # Define CLI
@@ -38,6 +38,8 @@ app.command(
     help="Create a predictor by compiling a Python function."
 )(compile_predictor)
 app.command(name="retrieve", help="Retrieve a predictor.")(retrieve_predictor)
+app.command(name="archive", help="Archive a predictor.")(archive_predictor)
+app.command(name="delete", help="Delete a predictor.")(delete_predictor)
 
 # Run
 if __name__ == "__main__":